requirements.txt
PyInstaller==6.0.0
PyQt5==5.15.9
Pillow==10.0.1
filetype==1.2.0